requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
dj-static==0.0.6
Django==5.1.4
django-bootstrap3==24.3
gunicorn==23.0.0
packaging==24.2
sqlparse==0.5.3
static3==0.7.0
typing_extensions==4.12.2
psycopg2>=2.6.1

pip~=24.2
zlib~=1.2.13
wheel~=0.44.0
openssl~=3.0.15
djangorestframework~=3.15.2